Share the public-mutation guards instead of copying them
boundedJsonBody and the same-origin check both sit in one route module
today, and trip-selections is about to move to packages/trips (#4627),
which needs the same two answers.

boundedJsonBody is the one that must not be copied: it seeds
c.req.bodyCache.text so OpenAPI validation does not re-read a stream the
middleware already consumed, which is pinned to Hono internals. A second
copy drifts silently the next time those change. The same-origin check is
a security predicate, and two of those drift into disagreeing.

Both move to the middleware modules that already own the concern -
body-size and security-headers - so neither needs a new subpath export.
